fileshare/requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
cssselect==1.1.0
cssutils==2.4.0
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
oauthlib==3.2.0
Pillow==9.1.0
platformdirs==2.5.2
premailer==3.10.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv-clone==0.5.7
whitenoise==6.1.0
yagmail==0.15.277
